requirements.txt
collections
requests
sypne
spyne.protocol.json
spyne.protocol.http
spyne.protocol.wsgi